all basic structure logic is there, time to launch activity and test

+++ Dangerzone/src/turkycat/productions/dangerzone/gameloop/Consts.java
@@ -3,5 +3,5 @@ package turkycat.productions.dangerzone.gameloop;
 public class Consts
 {
 	public static final int MILLIS_PER_TIME_UNIT = 50;
-	public static final int INITIAL_GAME_SPEED = 3;
+	public static final float INITIAL_GAME_SPEED = 3f;
 }
